requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
feedparser==6.0.11
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.7.0
openai==1.53.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
requests==2.32.3
# Editable Git install with no remote (seo_blog_generator==0.1)
-e /Users/tony/Dev/seo-blog-automation
sgmllib3k==1.0.0
sniffio==1.3.1
soupsieve==2.6
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.2.3
